Made main and player page to  be on one page
