+++ Views/Player.xaml
@@ -5,40 +5,5 @@
              Title="Player"
              BackgroundColor="#C2BBF0">
 
-    <VerticalStackLayout>
-        <Border x:Name="DisplayBorder"
-                Stroke="#F1E3F3"
-                StrokeThickness="5"
-                StrokeShape="RoundRectangle 40,40,40,40"
-                Background="#3590F3"
-                Padding="16,8"
-                HorizontalOptions="Center"
-                Margin="0,100,0,100">
-            <Label x:Name="Display"
-                   Text="20:00"
-                   VerticalOptions="Center"
-                   HorizontalOptions="Center"
-                   Margin="10,5,5,10"
-                   FontSize="130"
-                   FontFamily="consolas"
-                   TextColor="#F1E3F3"/>
-        </Border>
-
-        <HorizontalStackLayout HorizontalOptions="Center" VerticalOptions="Center">
-            <Button x:Name="stopButton"
-                    Text="⬛"
-                    Clicked="Stop"
-                    HeightRequest="100"
-                    WidthRequest="100"
-                    Margin="0,0,100,0"/>
-
-            <Button x:Name="PlayPauseButton"
-                    Text="II"
-                    FontSize="30"
-                    Clicked="PlayPause"
-                    HeightRequest="100"
-                    WidthRequest="100"
-                    Margin="100,0,0,0"/>
-        </HorizontalStackLayout>
-    </VerticalStackLayout>
+        
 </ContentPage>
